backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
requests
psycopg2-binary
gunicorn
whitenoise
mysql-connector-python
boto3
